# docs/openapi/api-gateway.json
{
  "components": {
    "schemas": {
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "format": "email",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "type": "object"
      },
      "TrebovanjeDetail": {
        "properties": {
          "broj_stavki": {
            "type": "integer"
          },
          "datum": {
            "format": "date-time",
            "type": "string"
          },
          "dokument_broj": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "magacin": {
            "type": "string"
          },
          "radnja": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "stavke": {
            "items": {
              "$ref": "#/components/schemas/TrebovanjeItemDetail"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TrebovanjeImportPayload": {
        "properties": {
          "datum": {
            "format": "date-time",
            "type": "string"
          },
          "dokument_broj": {
            "type": "string"
          },
          "magacin_naziv": {
            "type": "string"
          },
          "magacin_pantheon_id": {
            "type": "string"
          },
          "radnja_naziv": {
            "type": "string"
          },
          "radnja_pantheon_id": {
            "type": "string"
          },
          "stavke": {
            "items": {
              "properties": {
                "artikl_sifra": {
                  "type": "string"
                },
                "barkod": {
                  "type": "string"
                },
                "kolicina_trazena": {
                  "type": "number"
                },
                "naziv": {
                  "type": "string"
                }
              },
              "required": [
                "artikl_sifra",
                "naziv",
                "kolicina_trazena"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "dokument_broj",
          "datum",
          "magacin_pantheon_id",
          "radnja_pantheon_id",
          "stavke"
        ],
        "type": "object"
      },
      "TrebovanjeItemDetail": {
        "properties": {
          "artikl_sifra": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kolicina_trazena": {
            "type": "number"
          },
          "kolicina_uradjena": {
            "type": "number"
          },
          "naziv": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TrebovanjeListItem": {
        "properties": {
          "broj_stavki": {
            "type": "integer"
          },
          "datum": {
            "format": "date-time",
            "type": "string"
          },
          "dokument_broj": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "magacin": {
            "type": "string"
          },
          "radnja": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "ukupno_trazena": {
            "type": "number"
          },
          "ukupno_uradjena": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "dokument_broj"
        ],
        "type": "object"
      },
      "TrebovanjeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrebovanjeListItem"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Magacin API Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Token response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Login"
      }
    },
    "/api/auth/me": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Current user profile"
      }
    },
    "/api/catalog/articles": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "aktivan": {
                            "type": "boolean"
                          },
                          "barkodovi": {
                            "items": {
                              "properties": {
                                "is_primary": {
                                  "type": "boolean"
                                },
                                "value": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "value"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "jedinica_mjere": {
                            "type": "string"
                          },
                          "naziv": {
                            "type": "string"
                          },
                          "sifra": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "sifra",
                          "naziv",
                          "jedinica_mjere",
                          "aktivan",
                          "barkodovi"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "page_size": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total",
                    "page",
                    "page_size"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Catalog article list"
          }
        },
        "summary": "List catalog articles",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/catalog/articles/{article_id}": {
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "aktivan": {
                    "type": "boolean"
                  },
                  "barkodi": {
                    "items": {
                      "properties": {
                        "is_primary": {
                          "type": "boolean"
                        },
                        "value": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "value"
                      ],
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "jedinica_mjere": {
                    "type": "string"
                  },
                  "naziv": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "aktivan": {
                      "type": "boolean"
                    },
                    "barkodovi": {
                      "items": {
                        "properties": {
                          "is_primary": {
                            "type": "boolean"
                          },
                          "value": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "value"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "jedinica_mjere": {
                      "type": "string"
                    },
                    "naziv": {
                      "type": "string"
                    },
                    "sifra": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "sifra",
                    "naziv",
                    "jedinica_mjere",
                    "aktivan",
                    "barkodovi"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated article"
          }
        },
        "summary": "Update catalog article",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/catalog/sync": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "deactivate_missing": {
                    "type": "boolean"
                  },
                  "path": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "cached": {
                      "type": "boolean"
                    },
                    "created": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "deactivated": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "duration_ms": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "finished_at": {
                      "format": "date-time",
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "message": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "payload_hash": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "processed": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "source": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "started_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "updated": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "started_at",
                    "cached"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Sync executed"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "cached": {
                      "type": "boolean"
                    },
                    "created": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "deactivated": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "duration_ms": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "finished_at": {
                      "format": "date-time",
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "message": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "payload_hash": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "processed": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "source": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "started_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "updated": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "started_at",
                    "cached"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Sync skipped (cached)"
          }
        },
        "summary": "Trigger catalog sync",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/trebovanja": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "default": 20,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "magacin_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "radnja_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrebovanjeListResponse"
                }
              }
            },
            "description": "Paged result"
          }
        },
        "summary": "List trebovanja"
      }
    },
    "/api/trebovanja/import": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrebovanjeImportPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Import trebovanje"
      }
    },
    "/api/trebovanja/{trebovanje_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "trebovanje_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrebovanjeDetail"
                }
              }
            },
            "description": "Trebovanje detail"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Get trebovanje detail"
      }
    },
...
  }
}